requirements.txt
streamlit
openai
yfinance
anthropic
python-dotenv
requests
pandas
numpy
plotly
newsapi-python
google-genai
groq
